# apps/web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NovelOS API",
    "version": "3.8.0"
  },
  "paths": {
...
    "/api/capability-bindings/{capability}": {
      "put": {
        "tags": [
          "capability_bindings"
        ],
        "summary": "Upsert Capability Binding",
        "description": "upsert 一个 binding。body: ``{profile_ids: [...]}``。\n\n- ``profile_ids`` 非空；\n- 所有 profile_id 必须存在且 ``enabled=1``（失效的 profile 不允许出现在 binding 里——\n  否则 candidate 链会跳过它，违反「binding 显式声明」的语义）；\n- 未知 capability → 404；\n- 校验失败 → 422（detail 含具体原因）。",
        "operationId": "upsert_capability_binding_api_capability_bindings__capability__put",
        "parameters": [
          {
            "name": "capability",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Capability"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Upsert Capability Binding Api Capability Bindings  Capability  Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "capability_bindings"
        ],
        "summary": "Delete Capability Binding",
        "description": "解除 binding；不存在 → 404。",
        "operationId": "delete_capability_binding_api_capability_bindings__capability__delete",
        "parameters": [
          {
            "name": "capability",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Capability"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Capability Binding Api Capability Bindings  Capability  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}